harchi.app-main/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
django-filter==22.1
django-cors-headers==3.14.0 
drf-yasg
requests
beautifulsoup4
pandas
jdatetime
hijri-converter
geonamescache
PrayTimes
timezonefinder
hijridate